requirements.txt
streamlit
youtube-transcript-api
google-generativeai
yt-dlp-transcript